% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_plot.R
\name{convergence_plot}
\alias{convergence_plot}
\title{convergence_plot}
\usage{
convergence_plot(model, Xtest, ytest, burnin = 0)
}
\arguments{
\item{model}{list containing a model of class "hs_rulefit".}

\item{Xtest}{Out of bag sample to check error.}

\item{ytest}{response of test data.}

\item{burnin}{Number of samples disregarded as burnin.}
}
\description{
Can be used to check model convergence.
}
\details{
To check convergence on the Training data just use training X and y instead of Xtest and ytest.
}

