% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_hs.R
\name{importance_hs}
\alias{importance_hs}
\title{Most important Rules/terms}
\usage{
importance_hs(model, k = 10)
}
\arguments{
\item{model}{list containing a model of class "HorseRuleFit".}

\item{k}{number of most important rules to be shown in the table.}
}
\description{
Produces a table containing the most important Rules or linear terms
}

