% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruleheat.R
\name{ruleheat}
\alias{ruleheat}
\title{RuleHeatmap}
\usage{
ruleheat(model, k)
}
\arguments{
\item{model}{list containing a model of class "HorseRuleFit".}

\item{k}{number of most important rules to be shown in the RuleHeat plot.}
}
\description{
Produces a heatmap that allows to identify what observations are covered by the most important decision rules. Details can be found in Nalenz & Villani (2017).
}

