% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{Variable_importance}
\alias{Variable_importance}
\title{Variable Importance plot}
\usage{
Variable_importance(model, top = NULL, var_names = NULL)
}
\arguments{
\item{model}{list containing a model of class "hs_rulefit".}

\item{top}{If a integer number is given only shows the top most important variables in the plot}

\item{var_names}{optional vector with the variable names to be shown in plot.}
}
\description{
Creates a input variable importance plot
}

